Create a separate file
