+++ Source/CacheAware.swift
@@ -1,19 +1,5 @@
 import Foundation
 
-public class CacheTask {
-  private var block: dispatch_block_t
-
-  init(processing: () -> Void) {
-    block = dispatch_block_create(DISPATCH_BLOCK_INHERIT_QOS_CLASS) {
-      processing()
-    }
-  }
-
-  func cancel() {
-    dispatch_block_cancel(block)
-  }
-}
-
 public protocol CacheAware {
   var prefix: String { get }
   var path: String { get }
